// src/types/errorCodes.ts
export enum ErrorCode {
    // 400 BAD REQUEST
    EMPTY_BODY = "EMPTY_BODY",
    INVALID_CUSTOMER_ID = "INVALID_CUSTOMER_ID",
    INVALID_ACCOUNT_ID = "INVALID_ACCOUNT_ID",
    MISSING_REQUIRED_FIELDS = "MISSING_REQUIRED_FIELDS",
    DUPLICATE_ACCOUNT = "DUPLICATE_ACCOUNT",

    // 401 UNAUTHORIZED
    TOKEN_EXPIRED = "TOKEN_EXPIRED",

    // 404 NOT FOUND
    ACCOUNT_NOT_FOUND = "ACCOUNT_NOT_FOUND",

    // 409 CONFLICT
    USERNAME_ALREADY_EXISTS = "USERNAME_ALREADY_EXISTS",

    // 500 INTERNAL SERVER ERROR
    INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
}